Cambios de documentación

+++ src/main/java/es/uma/informatica/sii/helloworld/service/ListaCompraDBService.java
@@ -18,43 +18,52 @@ import es.uma.informatica.sii.helloworld.repositories.ListaCompraRepository;
 @Service
 @Transactional
 public class ListaCompraDBService {
-	
+
 	private ListaCompraRepository listaRepository;
 	private ItemRepository itemRepository;
-	
+
 	@Autowired
 	public ListaCompraDBService(ListaCompraRepository repo, ItemRepository itemRepository) {
 		this.listaRepository = repo;
 		this.itemRepository = itemRepository;
 	}
-	
+
 	public Optional<ListaCompra> listaCompraById(Long id) {
 		return listaRepository.findById(id);
 	}
-	
+
 	public List<Long> todasListascompras() {
 		return StreamSupport.stream(listaRepository.findAll().spliterator(),false)
 				.map(ListaCompra::getId).toList();
 	}
-	
+
 	public Long aniadirListaCompra(ListaCompra lc) {
 		listaRepository.save(lc);
 		return lc.getId();
 	}
-	
+
 	public void modificarListaCompra(ListaCompra lista) {
-		listaRepository.findById(lista.getId()).ifPresent(l->l.setNombre(lista.getNombre()));
+		if (listaRepository.existsById(lista.getId())) {
+			listaRepository.findById(lista.getId())
+			.ifPresent(l->l.setNombre(lista.getNombre()));
+		} else {
+			throw new NoEncontradoException();
+		}
 	}
-	
+
 	public void eliminarListaCompra(Long id) {
-		listaRepository.deleteById(id);
+		if (listaRepository.existsById(id)) {
+			listaRepository.deleteById(id);
+		} else {
+			throw new NoEncontradoException();
+		}
 	}
-	
+
 	public Long aniadirItem(Long idLista, Item item) {
 		var lista = listaRepository.findById(idLista);
 		if (lista.isPresent()) {
 			item = itemRepository.save(item);
-			
+
 			ListaCompra lc = lista.get();
 			lc.getItems().add(item);
 			return item.getId();			
@@ -62,18 +71,26 @@ public class ListaCompraDBService {
 			return null;
 		}
 	}
-	
+
 	public void modificarItem(Long idLista, Item item) {
-		itemRepository.save(item);
+		if (itemRepository.existsById(item.getId())) {
+			itemRepository.save(item);
+		} else {
+			throw new NoEncontradoException();
+		}
 	}
-	
+
 	public void eliminarItem(Long idLista, Long idItem) {
-		itemRepository.deleteById(idItem);
+		if (itemRepository.existsById(idItem)) {
+			itemRepository.deleteById(idItem);
+		} else {
+			throw new NoEncontradoException();
+		}
 	}
-	
+
 	public Optional<Item> getItem(Long idLista, Long idItem) {
 		return itemRepository.findById(idItem);
 	}
-	
+
 
 }
